Move surepetcare coordinator to separate module
